refactor code to simplify, improved class and package naming
