+++ backend/src/main/java/ca/bc/gov/hlth/accounttransfer/controller/AccountsController.java
@@ -2,12 +2,11 @@ package ca.bc.gov.hlth.accounttransfer.controller;
 
 import javax.validation.Valid;
 
-import ca.bc.gov.hlth.accounttransfer.applications.ClientsLookup;
-import ca.bc.gov.hlth.accounttransfer.applications.KeycloakClient;
-import ca.bc.gov.hlth.accounttransfer.model.rest.StatusEnum;
-import ca.bc.gov.hlth.accounttransfer.model.rest.accountTransfer.AccountResponse;
-import ca.bc.gov.hlth.accounttransfer.model.rest.ldap.LdapRequest;
-import ca.bc.gov.hlth.accounttransfer.model.rest.ldap.LdapResponse;
+import ca.bc.gov.hlth.accounttransfer.model.accountTransfer.AccountTransferResponse;
+import ca.bc.gov.hlth.accounttransfer.model.accountTransfer.StatusEnum;
+import ca.bc.gov.hlth.accounttransfer.util.keycloak.ClientsLookup;
+import ca.bc.gov.hlth.accounttransfer.model.ldap.LdapRequest;
+import ca.bc.gov.hlth.accounttransfer.model.ldap.LdapResponse;
 import ca.bc.gov.hlth.accounttransfer.service.KeycloakUserManagementService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,10 +18,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import ca.bc.gov.hlth.accounttransfer.model.rest.accountTransfer.AccountTransferRequest;
+import ca.bc.gov.hlth.accounttransfer.model.accountTransfer.AccountTransferRequest;
 import ca.bc.gov.hlth.accounttransfer.service.LdapService;
-
-import java.util.Arrays;
 
 /**
  * Handle requests related to Account Transfers.
@@ -36,6 +33,9 @@ public class AccountsController {
 	private static final String MSPDIRECT = "MSPDIRECT-SERVICE";
 	private static final String subClaim = "sub";
 
+	public AccountsController() {
+
+	}
 	@Autowired
 	private ClientsLookup clientsLookup;
 
@@ -53,36 +53,41 @@ public class AccountsController {
 	 * @return The result of the operation
 	 */
 	@PostMapping("/accountTransfer")
-	public ResponseEntity<AccountResponse> transferAccount(@Valid @RequestBody AccountTransferRequest accountTransferRequest) {
+	public ResponseEntity<AccountTransferResponse> transferAccount(@Valid @RequestBody AccountTransferRequest accountTransferRequest) {
 
 		Jwt authToken = (Jwt) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 		String userId = authToken.getClaimAsString(subClaim);
 
+		logger.debug("Attempting account transfer for User {} on Application {}", accountTransferRequest.getUsername(), accountTransferRequest.getApplication());
+
 		LdapRequest ldapRequest = new LdapRequest(accountTransferRequest.getUsername(), accountTransferRequest.getPassword());
 		LdapResponse ldapResponse = ldapService.checkLdapAccount(ldapRequest);
 
-		// todo if ldapResponse.username == '' -> error username doesn't exist
-		// todo if ldapResponse.authenticated == false -> error user is locked
-		// todo if ldapResponse.locked == true -> error account is locked
-		// todo if ldapResponse.mspDirectRole = '' -> error user doesn't actually have the role
+		// TODO if ldapResponse.username == '' -> error username doesn't exist
+		// TODO if ldapResponse.authenticated == false -> error user is locked
+		// TODO if ldapResponse.locked == true -> error account is locked
+		// TODO if ldapResponse.mspDirectRole = '' -> error user doesn't actually have the role
+		// TODO if ldapResponse.mspDirectRole is not supported in the new MSP Direct -> role is invalid
 
-		// Authentication success, account unlocked, user has a role for the application
+		logger.debug("User has role {}", ldapResponse.getMspDirectRole().toUpperCase());
+		// Authentication success, account unlocked, user has a role for the mspdirect
 		// Add the role to the user in Keycloak
 		if (accountTransferRequest.getApplication().equals(MSPDIRECT) && !ldapResponse.getMspDirectRole().equals("")) {
 
 			String idOfClient = clientsLookup.getClients().get(MSPDIRECT).getId();
+			// TODO VISARESIDENT role with return null here because it's now called VISARESIDENTS
 			Object roleRepresentation = clientsLookup.getClients().get(MSPDIRECT).getRoles().get(ldapResponse.getMspDirectRole().toUpperCase());
 			// Keycloak expects the representation of roles to add in an array
 			Object[] rolesToSend = { roleRepresentation };
 
 			keycloakUserManagementService.addRoleToUser(userId, idOfClient, rolesToSend);
 
-			AccountResponse response = new AccountResponse();
-			response.setStatus(StatusEnum.SUCCESS);
-			response.setMessage("Account transfer succeeded for user " + ldapResponse.getUsername() + " with application " + MSPDIRECT + " and role " + ldapResponse.getMspDirectRole());
-			ResponseEntity<AccountResponse> responseEntity = ResponseEntity.ok(response);
-			return responseEntity;
+			String responseMessage = "Account transfer succeeded for user " + accountTransferRequest.getUsername() + " for the application " + MSPDIRECT + " with the role " + ldapResponse.getMspDirectRole();
+			AccountTransferResponse response = new AccountTransferResponse(StatusEnum.SUCCESS, responseMessage, ldapResponse.getMspDirectRole().toUpperCase());
+
+			return ResponseEntity.ok(response);
 		}
+
 		return null;
