+++ backend/src/main/java/ca/bc/gov/hlth/accounttransfer/model/accountTransfer/AccountTransferRequest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.hlth.accounttransfer.model.rest.accountTransfer;
+package ca.bc.gov.hlth.accounttransfer.model.accountTransfer;
 
 public class AccountTransferRequest {
