--- backend/src/main/java/ca/bc/gov/hlth/accounttransfer/model/rest/BaseResponse.java
@@ -1,23 +1,0 @@
-package ca.bc.gov.hlth.accounttransfer.model.rest;
-
-public class BaseResponse {
-	protected StatusEnum status;
-	protected String message;
-
-	public StatusEnum getStatus() {
-		return status;
-	}
-
-	public void setStatus(StatusEnum status) {
-		this.status = status;
-	}
-
-	public String getMessage() {
-		return message;
-	}
-
-	public void setMessage(String message) {
-		this.message = message;
-	}
-
-}
