+++ backend/src/main/java/ca/bc/gov/hlth/accounttransfer/service/LdapService.java
@@ -1,9 +1,7 @@
 package ca.bc.gov.hlth.accounttransfer.service;
 
-import ca.bc.gov.hlth.accounttransfer.model.rest.ldap.LdapRequest;
-import ca.bc.gov.hlth.accounttransfer.model.rest.ldap.LdapResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import ca.bc.gov.hlth.accounttransfer.model.ldap.LdapRequest;
+import ca.bc.gov.hlth.accounttransfer.model.ldap.LdapResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
